requirements.txt
sanic==22.6.2
asyncpg==0.26.0
bcrypt==4.0.0
pyjwt==2.4.0